Fix formatting

+++ parafac2/normalize.py
@@ -9,11 +9,11 @@ from sklearn.utils.sparsefuncs import (
 
 
 def normalize_total(adata: anndata.AnnData):
-    counts_per_cell = np.array(adata.X.sum(axis=1)).flatten() # type: ignore
+    counts_per_cell = np.array(adata.X.sum(axis=1)).flatten()  # type: ignore
     cell_subset = counts_per_cell > 0
 
-    if issubclass(adata.X.dtype.type, (int, np.integer)): # type: ignore
-        adata.X = adata.X.astype(np.float32) # type: ignore
+    if issubclass(adata.X.dtype.type, (int, np.integer)):  # type: ignore
+        adata.X = adata.X.astype(np.float32)  # type: ignore
 
     counts_per_cell /= np.median(counts_per_cell[cell_subset])
